perf: fold Gemma (1+w) scale in x.dtype so vllm_c rms_norm fires
gemma_rms_norm folded weight.float()+1.0 producing an fp32 scale, which
fails the vllm_c impl's weight.dtype==x.dtype guard and silently fell
back to the decomposed native path (pow/mean/rsqrt/mul + fp16<->fp32
casts, ~5 kernels per call, 16987 calls per 1k/512 bench).

Validated on Qwen3.8-27B-AWQ-INT4 TP=2 gfx1030: greedy output unchanged
(8/8 probes), c=8 1k/512 wall 15.34s -> 14.92s.

+++ vllm/model_executor/layers/layernorm.py
@@ -360,7 +360,9 @@ def gemma_rms_norm(
     epsilon: float,
 ) -> torch.Tensor:
     """Gemma RMSNorm: x * (1+w) / rms. Opaque to inductor."""
-    scale = weight.float() + 1.0
+    # Fold in x.dtype: an fp32 scale fails the vllm_c dtype-match guard and
+    # silently falls back to the decomposed native path (~5 kernels/call).
+    scale = (weight.float() + 1.0).to(x.dtype)
     return ir.ops.rms_norm(x, scale, epsilon)
 
 
@@ -386,7 +388,7 @@ def gemma_fused_add_rms_norm(
     epsilon: float,
 ) -> tuple[torch.Tensor, torch.Tensor]:
     """Gemma fused residual+RMSNorm. Opaque to inductor."""
-    scale = weight.float() + 1.0
+    scale = (weight.float() + 1.0).to(x.dtype)
     return ir.ops.fused_add_rms_norm(x, residual, scale, epsilon)
